gemini-3-transcribe/requirements.txt
google-generativeai>=0.8.0
pydub>=0.25.1
yt-dlp>=2024.9.27
flask>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
